Two improvements to xml2pot
Reattach the attributes when the msgid of two different tags match.i.e.
<link linkend="Bit-Operations">Bit Operations</link>
and
<title>Bit Operations</title>

Do not kill a single tag at the end of a non tagged msgid, i.e.
Final step is to generate static &HTML; pages and upload to my web server for the enjoyment of friends and colleagues. &HTML; generation is described in <xref linkend="chp-generating-html"/>

Let's hope i did not break lost of other stuff :D

svn path=/trunk/KDE/kdesdk/poxml/; revision=1282429