switch from Antlr2 to gettext-po
Switch the Antlr2 build dependency of po2xml and swappo to gettext-po,
using it in them for reading .po files.

In the process, swappo has been rewritten to use gettext-po also for
outputting the resulting .po file, fixing the formatting and ordering
issues which were so far.

Theoretically there should be no behaviour changes in these tools.